+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.0.6
+
+- [增] `UploaderOptions.abortTransport` 自定义上传接口，当用户中止时回调
+
+## Bug Fixes
+
+- **Uploader** `auto:true` ([#8](https://github.com/cipchk/ngx-weui/issues/8))
+
+
 # 1.0.5
 
 - [增] 微信JS-SDK。
@@ -9,4 +18,4 @@
 
 ## Bug Fixes
 
-- **InfiniteLoader** 组件设置标签默认为 `display:block`。 ([#3](https://github.com/cipchk/vscode-cssrem/pull/3))
+- **InfiniteLoader** 组件设置标签默认为 `display:block`。 ([#3](https://github.com/cipchk/ngx-weui/issues/3))